# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tea Sequence API
  version: 1.0.0
  description: >-
    Stable v1 contract for Tea Sequence. The initial configuration response
    explicitly describes non-production demonstration mode.
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /api/v1/inquiries:
    post:
      operationId: createInquiry
      security: []
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/InquiryCreate'}}}
      responses:
        '201': {$ref: '#/components/responses/ApiSuccess'}
        '409': {$ref: '#/components/responses/ApiFailure'}
        '422': {$ref: '#/components/responses/ApiFailure'}
...
  /api/v1/admin/import-templates/{resource}:
    get:
      operationId: getImportTemplate
      security: [{AdminSessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/Resource'}, {name: format, in: query, required: true, schema: {type: string, enum: [csv, xlsx]}}]
      responses: {'200': {description: Import template file}, '503': {$ref: '#/components/responses/ApiFailure'}}
  /api/v1/admin/imports:
    post:
      operationId: createImport
      security: [{AdminSessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CsrfToken'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {multipart/form-data: {schema: {type: object, required: [file, resource, schema_version], properties: {file: {type: string, format: binary}, resource: {type: string}, schema_version: {type: string}}}}}}
      responses: {'202': {$ref: '#/components/responses/ApiSuccess'}}
  /api/v1/admin/imports/{import_id}:
    get:
      operationId: getImport
      security: [{AdminSessionCookie: []}]
      parameters: [{name: import_id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {$ref: '#/components/responses/ApiSuccess'}}
  /api/v1/admin/imports/{import_id}/errors:
    get:
      operationId: listImportErrors
      security: [{AdminSessionCookie: []}]
      parameters: [{name: import_id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {$ref: '#/components/responses/ApiSuccess'}}
  /api/v1/admin/imports/{import_id}/commit:
    post:
      operationId: commitImport
      security: [{AdminSessionCookie: []}]
      parameters: [{name: import_id, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/CsrfToken'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'202': {$ref: '#/components/responses/ApiSuccess'}, '409': {$ref: '#/components/responses/ApiFailure'}}
  /api/v1/admin/inquiries:
    get:
      operationId: listInquiries
      security: [{AdminSessionCookie: []}]
      responses: {'200': {$ref: '#/components/responses/ApiSuccess'}}
  /api/v1/admin/inquiries/{inquiry_id}:
    get:
      operationId: getInquiry
      security: [{AdminSessionCookie: []}]
      parameters: [{name: inquiry_id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {$ref: '#/components/responses/ApiSuccess'}}
    patch:
      operationId: patchInquiry
      security: [{AdminSessionCookie: []}]
      parameters: [{name: inquiry_id, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/CsrfToken'}, {$ref: '#/components/parameters/IfMatch'}]
      responses: {'200': {$ref: '#/components/responses/ApiSuccess'}, '409': {$ref: '#/components/responses/ApiFailure'}}
  /api/v1/admin/audit-logs:
    get:
      operationId: listAuditLogs
      security: [{AdminSessionCookie: []}]
      responses: {'200': {$ref: '#/components/responses/ApiSuccess'}, '403': {$ref: '#/components/responses/ApiFailure'}}
components:
  securitySchemes:
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: tea_sequence_session
      description: HttpOnly administrator session cookie for later admin operations.
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required by later cookie-authenticated state-changing operations.
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Required by later state-changing operations; must be a UUID.
      schema:
        type: string
        format: uuid
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string, pattern: '^"rv-[1-9][0-9]*"$'}
...
    Resource:
      name: resource
      in: path
      required: true
      schema: {type: string, enum: [teas, tea-items, effects, brewing-recipes, suppliers, supply-offers, sources]}
...
  responses:
    ApiSuccess:
      description: Standard success envelope; operation-specific data is defined by the prose contract.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/SuccessObject'}
    ApiFailure:
      description: Standard failure envelope; clients branch on error.code.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  schemas:
    SuccessObject:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: {type: object}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
...
    InquiryCreate:
      type: object
      additionalProperties: false
      required: [kind, need, contact, consent]
      properties:
        kind: {type: string, enum: [consultation, sample]}
        tea_id: {type: [string, 'null'], format: uuid}
        tea_item_id: {type: [string, 'null'], format: uuid}
        supply_offer_id: {type: [string, 'null'], format: uuid}
        need: {type: string, minLength: 1, maxLength: 1000}
        contact:
          type: object
          additionalProperties: false
          required: [channel, value]
          properties: {channel: {type: string, enum: [phone, email]}, value: {type: string, maxLength: 254}}
        consent:
          type: object
          additionalProperties: false
          required: [accepted, notice_version, purpose]
          properties: {accepted: {const: true}, notice_version: {type: string}, purpose: {const: inquiry_followup}}
...
    ResponseMeta:
      type: object
      additionalProperties: false
      required: [request_id, server_time]
      properties:
        request_id:
          type: string
          format: uuid
        server_time:
          type: string
          format: date-time
    ErrorDetail:
      type: object
      additionalProperties: true
      properties:
        field:
          type: string
        reason:
          type: string
    ApiErrorBody:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          $ref: '#/components/schemas/ApiErrorBody'
        meta:
          $ref: '#/components/schemas/ResponseMeta'